+++ tests/test_flask_pushjack.py
@@ -22,47 +22,59 @@ def app(request):
     return app
 
 
-@parametrize('extension_class,expected', [
-    (FlaskAPNS, {'APNS_ENABLED': True,
-                 'APNS_SANDBOX': False,
-                 'APNS_CERTIFICATE': None,
-                 'APNS_HOST': 'gateway.push.apple.com',
-                 'APNS_PORT': 2195,
-                 'APNS_FEEDBACK_HOST': 'feedback.push.apple.com',
-                 'APNS_FEEDBACK_PORT': 2196,
-                 'APNS_ERROR_TIMEOUT': 0.5,
-                 'APNS_DEFAULT_EXPIRATION_OFFSET': 60 * 60 * 24 * 30,
-                 'APNS_MAX_NOTIFICATION_SIZE': 2048}),
-    (FlaskAPNS, {'APNS_ENABLED': True,
-                 'APNS_SANDBOX': True,
-                 'APNS_CERTIFICATE': None,
-                 'APNS_HOST': 'gateway.sandbox.push.apple.com',
-                 'APNS_PORT': 2195,
-                 'APNS_FEEDBACK_HOST': 'feedback.sandbox.push.apple.com',
-                 'APNS_FEEDBACK_PORT': 2196,
-                 'APNS_ERROR_TIMEOUT': 0.5,
-                 'APNS_DEFAULT_EXPIRATION_OFFSET': 60 * 60 * 24 * 30,
-                 'APNS_MAX_NOTIFICATION_SIZE': 2048}),
-    (FlaskGCM, {'GCM_API_KEY': None,
-                'GCM_URL': 'https://android.googleapis.com/gcm/send',
-                'GCM_MAX_RECIPIENTS': 1000})
+@parametrize('extension_class,base_config,expected', [
+    (FlaskAPNS,
+     {'APNS_CERTIFICATE': None},
+     {'APNS_ENABLED': True,
+      'APNS_SANDBOX': False,
+      'APNS_CERTIFICATE': None,
+      'APNS_DEFAULT_ERROR_TIMEOUT': 10,
+      'APNS_DEFAULT_EXPIRATION_OFFSET': 60 * 60 * 24 * 30,
+      'APNS_DEFAULT_BATCH_SIZE': 100}),
+    (FlaskAPNS,
+     {'APNS_CERTIFICATE': None, 'APNS_SANDBOX': True},
+     {'APNS_ENABLED': True,
+      'APNS_SANDBOX': True,
+      'APNS_CERTIFICATE': None,
+      'APNS_DEFAULT_ERROR_TIMEOUT': 10,
+      'APNS_DEFAULT_EXPIRATION_OFFSET': 60 * 60 * 24 * 30,
+      'APNS_DEFAULT_BATCH_SIZE': 100}),
+    (FlaskGCM,
+     {'GCM_API_KEY': None},
+     {'GCM_API_KEY': None})
 ])
-def test_default_configuration(app, extension_class, expected):
-    app.config.update(expected)
+def test_default_configuration(app, extension_class, base_config, expected):
+    app.config.update(base_config)
     client = extension_class()
     client.init_app(app)
     assert issubset(expected, app.config)
 
 
-@parametrize('extension_class,method,send_object,args', [
-    (FlaskAPNS, 'send', 'pushjack.apns.send', ([], 'hello')),
-    (FlaskAPNS, 'send_bulk', 'pushjack.apns.send_bulk', ([], 'hello')),
-    (FlaskAPNS, 'get_expired_tokens', 'pushjack.apns.get_expired_tokens', ()),
-    (FlaskGCM, 'send', 'pushjack.gcm.send', ([], 'hello')),
-    (FlaskGCM, 'send_bulk', 'pushjack.gcm.send_bulk', ([], 'hello')),
+@parametrize('extension_class,config,method,send_object,args', [
+    (FlaskAPNS,
+     {'APNS_CERTIFICATE': None},
+     'send',
+     'pushjack.apns.APNSClient.send',
+     ([], 'hello')),
+    (FlaskAPNS,
+     {'APNS_CERTIFICATE': None},
+     'get_expired_tokens',
+     'pushjack.apns.APNSClient.get_expired_tokens',
+     ()),
+    (FlaskGCM,
+     {'GCM_API_KEY': None},
+     'send',
+     'pushjack.gcm.GCMClient.send',
+     ([], 'hello')),
 ])
-def test_client_methods(app, extension_class, method, send_object, args):
+def test_client_methods(app,
+                        extension_class,
+                        config,
+                        method,
+                        send_object,
+                        args):
     with mock.patch(send_object) as patched:
+        app.config.update(config)
         client = extension_class()
         client.init_app(app)
 
@@ -70,5 +82,5 @@ def test_client_methods(app, extension_class, method, send_object, args):
             getattr(client, method)(*args)
 
         assert patched.called
-        send_args = list(args) + [app.config]
+        send_args = list(args)
         patched.assert_called_with(*send_args)